+++ R/get_predictions_tobit.R
@@ -29,7 +29,7 @@ get_predictions.tobit <- function(model,
   tcrit <- stats::qt(ci, df = dof)
 
   # special handling for survival regression with type = "quantile"
-  if (is.null(type) || type != "quantile") {
+  if (is.null(type) || !type %in% c("quantile", "uquantile")) {
     type <- "lp"
   }
   if (type == "quantile") {
@@ -43,6 +43,34 @@ get_predictions.tobit <- function(model,
     se.fit = se,
     ...
   )
+
+  # we need to shape data into long-format when type = "quantile",
+  # but only when it's a matrix - for median time, when argument `p` is also
+  # given, we don't have a matrix
+  if (type %in% c("quantile", "uquantile") && (is.matrix(prdat) || (!is.null(prdat$fit) && is.matrix(prdat$fit)))) { # nolint
+    ## FIXME: check matrix structure
+    # for type = "quantile", we get a matrix of predictions, with two columns
+    # (or maybe even more columns, one per status? need to check!)
+    # we now duplicate the data grid and add the two status options as
+    # "response" column to the data gridd
+    data_grid <- rbind(data_grid, data_grid)
+    data_grid$response.level <- 2
+    data_grid$response.level[1:(nrow(data_grid) / 2)] <- 1
+    # if SE are requested, we need to gather multiple columns
+    if (se) {
+      prdat <- .multiple_gather(
+        as.data.frame(prdat),
+        names_to = "status",
+        values_to = c("predicted", "se"),
+        columns = list(1:2, 3:4)
+      )
+      prdat <- list(fit = prdat$predicted, se.fit = prdat$se)
+    } else {
+      prdat <- .gather(as.data.frame(prdat), "status", "predicted")$predicted
+    }
+    # we now have "prdat" in the same structure as for other types, so we
+    # can proceed as usual from here...
+  }
 
   # did user request standard errors? if yes, compute CI
   if (se) {